posts/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==4.0.0
pika==1.3.2
